Move save_metadata from parser.py to Simulation object
* Collect saving of metadata in a single place
* Switch from JSON to YAML
