+++ ctwrap/__init__.py
@@ -1,6 +1,6 @@
 """Package exports."""
 from .simulation import Simulation, SimulationHandler
-from .parser import Parser, save_metadata
+from .parser import Parser
 from . import modules
 
 from ._version import __version__, __author__
